+++ Changelog.md
@@ -1,3 +1,13 @@
+## 0.5.0
+
+## ElectronNET.Core
+
+- Updated internal facade from `SocketIoFacade` to `ISocketConnection` (#1037)
+- Fixed `HasShadow` is not `true` by default (#1049) @AeonSake
+- Fixed missing `ReadAllLines` in MSBuild execution (#1035)
+- Fixed serialization of release notes for the updater (#1041)
+- Added automatic port selection and secured IPC communication (#1038)
+
 # 0.4.1
 
 ## ElectronNET.Core